pre-XP headers cleaned
git-svn-id: http://svn.miranda-ng.org/main/trunk@6183 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/FavContacts/src/csocket.cpp
@@ -7,7 +7,8 @@ int CSocket::Recv(char *buf, int count)
 
 int CSocket::Send(char *buf, int count)
 {
-	if (count < 0) count = strlen(buf);
+	if (count < 0)
+		count = (int)strlen(buf);
 	return send(m_socket, buf, count, 0);
 }
